Lots & lots of python refactoring
