+++ conf/modules.config
@@ -1591,19 +1591,6 @@ process {
             ]
         }
 
-        withName: CAT_CAT_QC {
-            ext.prefix  = { "${meta.id}_cat" }
-            publishDir = [
-                enabled: false
-            ]
-        }
-
-        withName: CAT_CAT_MMSEQS {
-            publishDir = [
-                enabled: false
-            ]
-        }
-
         withName: CHECKV_DOWNLOADDATABASE {
             publishDir = [
                 path: { "${params.outdir}/databases/" },
@@ -1658,9 +1645,8 @@ process {
 
         withName: BLASTN_QC {
             ext.args = '-max_target_seqs 5 -outfmt "6 qseqid sseqid stitle pident qlen slen length mismatch gapopen qstart qend sstart send evalue bitscore"' // don't change outfmt
-            ext.prefix = { "${meta.id}_${meta.step}" }
             publishDir = [
-                path: { "${params.outdir}/consensus/quality_control/blast/${meta.sample}/${meta.step}" },
+                path: { "${params.outdir}/consensus/quality_control/blast/" },
                 mode: params.publish_dir_mode,
                 pattern: "*.txt",
                 saveAs: { filename -> params.prefix || params.global_prefix  ? "${params.global_prefix}-$filename" : filename }
